update .h file

+++ get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: carlaugu <carlaugu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/06 15:57:51 by carlaugu          #+#    #+#             */
-/*   Updated: 2024/11/09 16:40:58 by carlaugu         ###   ########.fr       */
+/*   Updated: 2024/11/11 20:55:41 by carlaugu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@
 char	*get_next_line(int fd);
 void	*ft_calloc(size_t n, size_t size);
 int		ft_newline(char *read);
-int		ft_strlen(char *s);
+size_t		ft_strlen(char *s);
 char	*ft_first_check(int fd);
 char	*ft_clean_all(char *read, char *save);
 char	*ft_create_save(char *read, char *save);
